+++ operator/internal/resources/aisauth/configmap_test.go
@@ -25,7 +25,9 @@ var _ = Describe("ConfigMap", func() {
 			},
 			Spec: authv1alpha1.AIStoreAuthSpec{
 				Deployment: authv1alpha1.DeploymentSpec{
-					Image: "docker.io/aistorage/authn:v4.5",
+					Container: authv1alpha1.ContainerSpec{
+						Image: "docker.io/aistorage/authn:v4.5",
+					},
 				},
 			},
 		}