+++ operator/internal/resources/aisauth/deployment.go
@@ -50,8 +50,10 @@ func NewDeployment(authn *authv1alpha1.AIStoreAuth) (*appsv1ac.DeploymentApplyCo
 		return nil, err
 	}
 	checksum := sha256.Sum256([]byte(authnJSON))
-	spec := &authn.Spec.Deployment
-	container := newContainer(authn, spec)
+	container, err := newContainer(authn, &authn.Spec.Deployment.Container)
+	if err != nil {
+		return nil, err
+	}
 	podSpec := newPodSpec(authn, container)
 
 	return appsv1ac.Deployment(DeploymentName(authn), authn.Namespace).
@@ -72,8 +74,8 @@ func NewDeployment(authn *authv1alpha1.AIStoreAuth) (*appsv1ac.DeploymentApplyCo
 
 func newContainer(
 	authn *authv1alpha1.AIStoreAuth,
-	spec *authv1alpha1.DeploymentSpec,
-) *corev1ac.ContainerApplyConfiguration {
+	spec *authv1alpha1.ContainerSpec,
+) (*corev1ac.ContainerApplyConfiguration, error) {
 	container := corev1ac.Container().
 		WithName(containerName).
 		WithImage(spec.Image).
@@ -86,7 +88,35 @@ func newContainer(
 	if spec.ImagePullPolicy != "" {
 		container.WithImagePullPolicy(spec.ImagePullPolicy)
 	}
-	return container
+	if spec.Resources != nil {
+		resources, err := toApplyConfiguration[*corev1ac.ResourceRequirementsApplyConfiguration](spec.Resources)
+		if err != nil {
+			return nil, err
+		}
+		container.WithResources(resources)
+	}
+	if spec.SecurityContext != nil {
+		securityContext, err := toApplyConfiguration[*corev1ac.SecurityContextApplyConfiguration](spec.SecurityContext)
+		if err != nil {
+			return nil, err
+		}
+		container.WithSecurityContext(securityContext)
+	}
+	if spec.LivenessProbe != nil {
+		livenessProbe, err := toApplyConfiguration[*corev1ac.ProbeApplyConfiguration](spec.LivenessProbe)
+		if err != nil {
+			return nil, err
+		}
+		container.WithLivenessProbe(livenessProbe)
+	}
+	if spec.ReadinessProbe != nil {
+		readinessProbe, err := toApplyConfiguration[*corev1ac.ProbeApplyConfiguration](spec.ReadinessProbe)
+		if err != nil {
+			return nil, err
+		}
+		container.WithReadinessProbe(readinessProbe)
+	}
+	return container, nil
 }
 
 func newPodSpec(