+++ operator/internal/resources/aisauth/env_test.go
@@ -23,7 +23,9 @@ var _ = Describe("Secret environment", func() {
 			},
 			Spec: authv1alpha1.AIStoreAuthSpec{
 				Deployment: authv1alpha1.DeploymentSpec{
-					Image: "docker.io/aistorage/authn:v4.8",
+					Container: authv1alpha1.ContainerSpec{
+						Image: "docker.io/aistorage/authn:v4.8",
+					},
 				},
 			},
 		}