+++ operator/internal/resources/aisauth/persistence_test.go
@@ -27,7 +27,9 @@ var _ = Describe("Persistence", func() {
 			},
 			Spec: authv1alpha1.AIStoreAuthSpec{
 				Deployment: authv1alpha1.DeploymentSpec{
-					Image: "docker.io/aistorage/authn:v4.5",
+					Container: authv1alpha1.ContainerSpec{
+						Image: "docker.io/aistorage/authn:v4.5",
+					},
 				},
 			},
 		}